Refactor: make python code modular
