+++ utils.py
@@ -1,0 +1,2 @@
+BUYER = 1
+SELLER = 2
